Adding shortcode-reversals.php mu-plugin
On WPCOM, we are helping users pasting HTML by reversing the code and turning that into shortcodes. The code is based on http://mdawaffepresents.wordpress.com/?p=36

This commmit contains bare-bones of the code in order to help developers developing for WordPress.com VIP platform debug issues related to the reversal process.